chore: rename android package name following app id name
