+++ android/app/src/main/java/eu/forkbomb/didroom/MainActivity.java
@@ -1,4 +1,4 @@
-package didroom.forkbomb.eu;
+package eu.forkbomb.didroom;
 
 import android.os.Bundle;
