+++ android/app/src/main/java/eu/forkbomb/didroom/TEE.java
@@ -1,4 +1,4 @@
-package didroom.forkbomb.eu;
+package eu.forkbomb.didroom;
 
 import android.os.Build;
 import android.security.keystore.KeyGenParameterSpec;
@@ -14,7 +14,6 @@ import com.getcapacitor.annotation.CapacitorPlugin;
 
 import java.nio.ByteBuffer;
 import java.security.KeyStore;
-import java.security.SecureRandom;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.Enumeration;
@@ -23,7 +22,6 @@ import javax.crypto.Cipher;
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.GCMParameterSpec;
-import javax.crypto.spec.IvParameterSpec;
 
 @CapacitorPlugin(name = "TEE")
 public class TEE extends Plugin {
